Quick add a Work Item Type(almighty/almighty-core#99 UI Issue 3). (#23)
When viewing a list of work items, the top row will be used to quickly add a work item.
When the item name is entered the add button (+) gets enabled. After adding, the list of work items are displayed in reversed order so that the item that was added is displayed at the top.