fix(codebase-sort): fixing codebase sorting issue
After the codebases are fetched, an empty object is added to the beginning of the codebases array
for table header display purposes. When sorting happens, the sort function is trying compare
properties on the object and the empty object added at the beginning obviously does not have any
properties. So if the object is empty, skip it and return 0

1456