fix(title): remove duplicate title in wizard
remove the duplicate 'Create Space' title from the Space Wizard when creating a new Space from a user's profile. As the Space Creator component contains the title, it does not need to be passed by the page to the  modal itself.

fixes https://github.com/fabric8-ui/fabric8-ui/issues/1627