fix(work-item-widget): Fix specificity of .combobox selector override
The combobox selector was recently overridden for work-item-widget. This uses the ‘top’ property to position the select element absolutely. As a result, the select element used on the codebases page is incorrectly positioned at the top of its container. The codebases select element should appear in each row of the list component, not in the header.