fix(tooltip): update tooltip location to prevent scrollbar
When hovering over the '+' button, the tooltip was set to be placed to the right of the icon. With the icon set to the right, there was no room, causing the scroll and jumping. This PR sets the tooltip placement to the left of the icon, preventing the offscreen issues.

fixes https://github.com/openshiftio/openshift.io/issues/1181