fix(package-lock): update fabric8-planner, ngx-forge, and patternfly-ng (#3069)
Currently, prod-preview appears to have a mismatch of versions. The console errors suggest that an older version of ngx-forge/patternfly-ng is in use because it refers to unavailable code. As a solution, I've regenerated the package locks for ngx-forge, fabric8-planner, and fabric8-ui using the latest patternfly-ng. I've also ensured that patternfly-ng is compatible with the older ngx-forge code, just in case.

Fixes https://github.com/openshiftio/openshift.io/issues/3833