feat(roles): add feature flags on settings icon and collaborators page (#3434)
Planner Story - https://openshift.io/openshiftio/Openshift_io/plan/detail/1211

This PR adds feature flag to the logic to hiding settings icon in the header and on the collaborators page inside settings.

The header component is rendered by using MenusService to determine if the menu item needs to be rendered or not.
MenusService checks for feature enabled menu items and send it to component level where the logic to hide something based on user is owner or admin is decided.
This is why its tricky to introduce feature flag at the header level based on the user permission as well as the feature enablement.
For achieving this we had to introduce on more property called subFeature in the MenuItem type interface. Based on the feature property menu item is deleted from the list using menu.service. After that the component uses subFeature to further decide which logic to use based on faeture flag along with user permission.
Add collaborators page under feature flag.
Feature flag for both the pages are set to beta as planner is also set at same level.